Refactor visibility tracking for contexts as a separate class.
