+++ .rubocop.yml
@@ -59,3 +59,7 @@ Style/Documentation:
     - 'lib/reek/ast/sexp_extensions/send.rb'
     - 'lib/reek/ast/sexp_extensions/super.rb'
     - 'lib/reek/ast/sexp_extensions/variables.rb'
+
+Style/AccessorMethodName:
+  Exclude:
+    - 'lib/reek/context/visibility_tracker.rb'
